Address PR #4 review: bf16 storage, warp mask fix, head_dim assert, sync check
- CUDA kernels now use __nv_bfloat16 storage for Q/K/V/out/attn_weights
  with f32 compute (matching production FlashAttention-style precision)
- Gradients (d_attn_out, d_q, d_k, d_v) remain f32 throughout
- Replace 0xFFFFFFFF warp mask with __activemask() in both kernels
- Add head_dim <= 32 runtime assertion in dispatch (warp size constraint)
- Check cudaDeviceSynchronize return code after forward and backward
- Add DevBuf16 RAII wrapper and f32<->bf16 conversion helpers in dispatch
- Widen FD tolerances for bf16 path (abs_threshold 5e-3, FD_TOL 20%)
- All tests pass: 36/36 non-CUDA, 47/47 CUDA, zero warnings

+++ core/src/cuda_ffi.rs
@@ -4,38 +4,43 @@
 // linked by the `cc` crate at build time. They are opaque to Enzyme
 // because nvcc produces SASS/PTX, not LLVM IR.
 //
+// bf16 pointers use *const u16 / *mut u16 in Rust FFI, corresponding
+// to __nv_bfloat16* on the CUDA side (both are 16-bit types).
+// Gradient pointers (d_attn_out, d_q, d_k, d_v) remain f32.
+//
 // Only available when compiled with `--features cuda`.
 
 extern "C" {
-    /// CUDA SWA forward kernel.
+    /// CUDA SWA forward kernel (bf16 storage, f32 compute).
     ///
     /// All pointers must be device memory (allocated via cudaMalloc).
     /// Layout matches the Rust reference `swa::swa_forward()`:
-    ///   q, k, v:       [seq_len, num_heads * head_dim] row-major
-    ///   out:            [seq_len, num_heads * head_dim] row-major
-    ///   attn_weights:   [num_heads, seq_len, window_size]
+    ///   q, k, v:       [seq_len, num_heads * head_dim] row-major (bf16)
+    ///   out:            [seq_len, num_heads * head_dim] row-major (bf16)
+    ///   attn_weights:   [num_heads, seq_len, window_size] (bf16)
     pub(crate) fn swa_forward_f32_cuda(
-        q: *const f32,
-        k: *const f32,
-        v: *const f32,
-        out: *mut f32,
-        attn_weights: *mut f32,
+        q: *const u16,
+        k: *const u16,
+        v: *const u16,
+        out: *mut u16,
+        attn_weights: *mut u16,
         seq_len: i32,
         num_heads: i32,
         head_dim: i32,
         window_size: i32,
     );
 
-    /// CUDA SWA backward kernel.
+    /// CUDA SWA backward kernel (bf16 inputs, f32 gradients).
     ///
     /// Computes dQ, dK, dV from upstream d_attn_out and cached attn_weights.
     /// Layout matches `swa::swa_backward_rust()`.
+    /// Q/K/V/attn_weights are bf16; d_attn_out/dQ/dK/dV are f32.
     /// dQ, dK, dV must be pre-zeroed by the caller.
     pub(crate) fn swa_backward_f32_cuda(
-        q: *const f32,
-        k: *const f32,
-        v: *const f32,
-        attn_weights: *const f32,
+        q: *const u16,
+        k: *const u16,
+        v: *const u16,
+        attn_weights: *const u16,
         d_attn_out: *const f32,
         d_q: *mut f32,
         d_k: *mut f32,
